+++ src/contacts/users/index.ts
@@ -1,4 +1,3 @@
-export * from "./auth-scope.enum";
 export * from "./commander.class";
 export * from "./ext-attr.interface";
 export * from "./user.class";