+++ src/core/http-helper.ts
@@ -8,7 +8,12 @@ const checkResponse = (res: AxiosResponse) => {
 
 const handleHttpError = (err: any, url: string, method: string, data?: any) => {
   throw Error(
-    `HTTP_ERROR_FROM_WECHAT:${{ message: err.message, url, method, data }}`
+    `HTTP_ERROR_FROM_WECHAT:${JSON.stringify({
+      message: err.message,
+      url,
+      method,
+      data
+    })}`
   );
 };