+++ src/customized-apps/index.ts
@@ -1,3 +1,4 @@
+export * from "./auth-scope.enum";
 export * from "./customized-app-config.interface";
 export * from "./customized-app-simple.class";
 export * from "./customized-app.class";